Move all the basic cmd functions out to separate files from the cmdsets
